requirements.txt
-e ./python_withings_api
arrow==1.3.0
chardet==5.2.0
contourpy==1.1.1
cryptography
cycler==0.12.1
DateTime==5.4
django_extensions
django-cors-headers
django-filter
Django==3.2.23
djangorestframework
djangorestframework-simplejwt
fonttools==4.50.0
gunicorn
importlib-resources==6.3.1
kiwisolver==1.4.5
matplotlib==3.7.5
numpy
oauthlib==3.2.2
packaging==24.0
pandas==2.0.3
pillow==10.2.0
psycopg2
psycopg2-binary
pyparsing==3.1.2
PyPDF2
python-dateutil==2.9.0.post0
python-decouple
python-decouple 
pytz==2024.1
reportlab
scipy
seaborn==0.13.2
six==1.16.0
types-python-dateutil==2.9.0.20240316
tzdata==2024.1
zipp==3.18.1
zope.interface==6.2